requirements.txt
streamlit,
sentence_transformers,
faiss-cpu,
transformers,
torch